backend/requirements.txt
django==4.1.4
djangorestframework==3.14.0
Pillow==9.4.0
pytest==7.2.0
pytest_django==4.5.2
pytest_pythonpath==0.7.4
django-filter==22.1
